Make the Windows build fail the run when a command fails
GitHub Actions does not by default fail builds in PowerShell when a
command fails. Another bug was that the custom build has to run in its
own directory that is not a sub directory of ExifReader for it to
expose all errors.

Then fix the error that appears: a missing dependency for
terser-webpack-plugin.

#89